+++ src/app/Modules/Language/FillLyric/Commands/Cell/DeleteCellCmd.h
@@ -13,7 +13,7 @@ namespace FillLyric {
 
     private:
         CellList *m_list;
-        qlonglong m_index;
+        int m_index;
         LyricCell *m_cell;
     };
